Move js to src dir
