+++ test/bundled.js
@@ -1,4 +1,4 @@
-var Bundled = require('../bundled')
+var Bundled = require('../src/bundled')
 
 var should = require('should');
 
@@ -6,7 +6,7 @@ describe('Bundled', function() {
   it('loadEvents', function() {
     should.equal(Bundled._loadEvents().length, 48)
 
-    days = Bundled._loadEvents()[0].days()
+    var days = Bundled._loadEvents()[0].days()
     should.equal(days.length, 3);
     should(days[0]).be.a.Date();
   })
